Separates orc8r api client from rest
