+++ orchestrator-bundle/orc8r-bundle/tests/integration/test_integration.py
@@ -2,55 +2,17 @@
 # Copyright 2021 Canonical Ltd.
 # See LICENSE file for licensing details.
 
-from contextlib import contextmanager
-from pathlib import Path
-from tempfile import NamedTemporaryFile
-from typing import Iterator, Tuple
+
+from typing import Tuple
 
 import jinja2
 import pytest
-import requests  # type: ignore[import]
-import urllib3  # type: ignore[import]
-from cryptography.hazmat.primitives.serialization import (
-    Encoding,
-    NoEncryption,
-    PrivateFormat,
-)
-from cryptography.hazmat.primitives.serialization.pkcs12 import (
-    load_key_and_certificates,
-)
 from pytest_operator.plugin import OpsTest
 from python_hosts import Hosts, HostsEntry  # type: ignore[import]
 
-urllib3.disable_warnings()
+from tests.integration.orchestrator import Orc8r
 
 DOMAIN = "pizza.com"
-
-
-@contextmanager
-def pfx_to_pem(pfx_path: str, pfx_password: str) -> Iterator[str]:
-    """Decrypts the .pfx file to be used with requests.
-
-    Args:
-        pfx_path: PFX file path
-        pfx_password: PFX file password
-
-    Returns:
-        str: Temporary pem file path
-    """
-    pfx_file = Path(pfx_path).read_bytes()
-    private_key, certificate, _ = load_key_and_certificates(
-        pfx_file, pfx_password.encode("utf-8"), None
-    )
-    if not private_key or not certificate:
-        raise RuntimeError("Could not load pfx package")
-    with NamedTemporaryFile(suffix=".pem") as t_pem:
-        with open(t_pem.name, "wb") as pem_file:
-            pem_file.write(
-                private_key.private_bytes(Encoding.PEM, PrivateFormat.PKCS8, NoEncryption())
-            )
-            pem_file.write(certificate.public_bytes(Encoding.PEM))
-        yield t_pem.name
 
 
 def update_etc_host(
@@ -154,21 +116,6 @@ async def get_pfx_package(ops_test: OpsTest) -> str:
     if retcode != 0:
         raise RuntimeError(f"Error: {stderr}")
     return export_path
-
-
-class Orc8r:
-    """Class that represents a Magma Orchestrator Instance"""
-
-    def __init__(self, url: str, admin_operator_pfx_path: str, admin_operator_pfx_password: str):
-        self.url = url
-        self.admin_operator_pfx_path = admin_operator_pfx_path
-        self.admin_operator_pfx_password = admin_operator_pfx_password
-
-    def get(self, endpoint: str) -> requests.Response:
-        with pfx_to_pem(self.admin_operator_pfx_path, self.admin_operator_pfx_password) as cert:
-            response = requests.get(url=f"{self.url}{endpoint}", cert=cert, verify=False)
-            response.raise_for_status()
-            return response
 
 
 class TestOrc8rBundle:
